third_party/blink/web_tests/external/wpt/scroll-animations/css/animation-timeline-named-scroll-progress-timeline.tentative-expected.txt

This is a testharness.js-based test.
PASS scroll-timeline-name is referenceable in animation-timeline on the declaring element itself
PASS scroll-timeline-name is referenceable in animation-timeline on that element's descendants
FAIL scroll-timeline-name is not referenceable in animation-timeline on that element's siblings assert_equals: Animation with unknown timeline name holds current time at zero expected "50px" but got "100px"
PASS scroll-timeline-name on an element which is not a scroll-container
PASS Change in scroll-timeline-name to match animation timeline updates animation.
PASS Change in scroll-timeline-name to no longer match animation timeline updates animation.
PASS Timeline lookup updates candidate when closer match available.
PASS Timeline lookup updates candidate when match becomes available.
PASS scroll-timeline-axis is block
PASS scroll-timeline-axis is inline
PASS scroll-timeline-axis is horizontal
PASS scroll-timeline-axis is vertical
PASS scroll-timeline-axis is mutated
Harness: the test ran to completion.
